add hack that imports dynamic symbols with global export for Tcl8.6 and up
As of Tcl8.6 dynamically loaded objects such as the compiled code of
the TclMPI module are loaded into an anonymous namespace and thus render
all symbols from linked library invisible to later loaded dynamic
objects used in that library. The load command has a new flag -global
that restores the old behavior. To stay backward compatible we need to
add a test for whether the Tcl version is new enough.

this closes #27